% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetterDropNA.R
\name{BetterDropNA}
\alias{BetterDropNA}
\title{Drop rows where every column specified is NA}
\usage{
BetterDropNA(dat, checkCols)
}
\arguments{
\item{dat}{the input data frame}

\item{checkCols}{columns to check where data is missing}
}
\description{
pretty self-explanatory
}
\examples{
# newdat <- BetterDropNA(dat, checkCols=c("column1", "column3"))
}
