% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToTable.R
\name{ConvertToTable}
\alias{ConvertToTable}
\title{Convert list-form model outputs into a table}
\usage{
ConvertToTable(obj, flip_dir = FALSE)
}
\arguments{
\item{obj}{any list-output from one of the many functions in this package that generates model outputs}
}
\description{
Convert list-form model outputs into a table
}
\examples{
out <- DichotTest(data)
table_out <- bind_rows(lapply(out, ConvertToTable))
# this converts the coefficients into odds ratios
table_out[,c(1,5,6)] <- lapply(table_out[,c(1,5,6)], exp)

# if running tests where multilevel == TRUE, table formatting should be done like this:
multilevel_stess_correct <- QualitativeStatistics(data=merged_dat, id_var='MRN', group_var=ANALYZE_VAR,
                                                  tst_vars=c('n_pressors', 'EEG'), multilevel=T,
                                                  test_use='logistic_regress', correct_var='STESS_score')

multilevel_out_stess_lst <- lapply(multilevel_stess_correct, function(x) lapply(x, function(y) ConvertToTable(y)))
multilevel_out_stess_dfs <- lapply(multilevel_out_stess_lst, bind_rows)
multilevel_out_stess_dfs <- lapply(multilevel_out_stess_dfs, function(x) {
  x[,c(1,5,6)] <- lapply(x[,c(1,5,6)], exp)
  return(x)
})
}
