% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeCol.R
\name{createTimeCol}
\alias{createTimeCol}
\title{create uniformity across the datetime columns for all the redcap databases to load.
this function is meant to be used internally by loadRedcap}
\usage{
createTimeCol(x)
}
\arguments{
\item{x}{a list of dataframes to modify the datetime column for}
}
\description{
create uniformity across the datetime columns for all the redcap databases to load.
this function is meant to be used internally by loadRedcap
}
