% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DichotTest.R
\name{DichotTest}
\alias{DichotTest}
\title{Test for differences in dichotomized numeric variables}
\usage{
DichotTest(data, group_var, tst_vars, GCS_compare = 7,
  check_n_percents = FALSE)
}
\arguments{
\item{data}{the input data frame}

\item{group_var}{the variable to group by}

\item{tst_vars}{a vector of variables to dichotomize by their median and test}

\item{GCS_compare}{if parameter happens to be the Glasgow Coma Scale, then dichotomize by > 7 since that's apparently already an agreed upon split in the literature. Defaults to 7}

\item{check_n_percents}{set to TRUE if wish to also obtain a count table with percentages. Defaults to FALSE.}
}
\description{
Dichotomize a variable by > or <= the median value and then perform Fisher's exact test on the resulting contingency table between a grouping variable.
This function was formally named OrdinalVarTest.
}
\examples{
out <- DichotTest(data=df, group_var="Included_in_Study", tst_vars=c("Admission_GCS", "Age", "ICU_Stay_Len"), check_n_percents=TRUE)
# see n(\%) tables
out$Age$n_percent
}
