% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallDevPkgs.R
\name{InstallDevPkgs}
\alias{InstallDevPkgs}
\title{Install Development packages}
\usage{
InstallDevPkgs()
}
\value{
(Invisibly) A data frame with information about the installed
package(s).
}
\description{
This calls \code{\link[=PkgInstall]{PkgInstall()}} updating common packages used in
development.
}
