% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevtools-package.R
\docType{package}
\name{kevtools-package}
\alias{kevtools}
\alias{kevtools-package}
\title{kevtools: Personal tools to make using R easier}
\description{
Short functions and tools to make using R and maintaining packages easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kevanness/kevtools}
  \item Report bugs at \url{https://github.com/Kevanness/kevtools/issues}
}

}
\author{
\strong{Maintainer}: Kevin van Ness \email{kevanness@gmail.com}

}
\keyword{internal}
