% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAll.R
\name{LoadAll}
\alias{LoadAll}
\title{Load complete package}
\usage{
LoadAll(package, repoPath = getOption("repoPath"))
}
\arguments{
\item{package}{Folder/path the package is in, starting from the repoPath.}

\item{repoPath}{Path to the directory.}
}
\description{
This calls [pkgload::load_all()] which loads a package. This
  splits the directory and package which can be useful for working on
  multiple packages without switching working directories or providing the
  entire path.
}
