% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRedcap.R
\name{loadRedcap}
\alias{loadRedcap}
\title{source in the latest version of data from a redcap database}
\usage{
loadRedcap(path, rcids)
}
\arguments{
\item{path}{a given path, or vector of paths, conatining a specific file or files to source}

\item{rcids}{a given path, or vector of paths, to the files containing the redcap ID to mrn link}
}
\description{
source in the latest version of data from a redcap database
}
