% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalInstallRevDevDeps.R
\name{LocalInstallRevDevDeps}
\alias{LocalInstallRevDevDeps}
\title{Install Reverse Development Dependencies}
\usage{
LocalInstallRevDevDeps(root = ".", ask = interactive())
}
\arguments{
\item{root}{Path to the package tree.}

\item{ask}{Whether to ask for confirmation when installing a different
version of a package that is already installed. Installations that only
add new packages never require confirmation.}
}
\value{
\code{root} invisibly.
}
\description{
This function identifies the local development dependencies of a package and their reverse dependencies, then installs all of them.
}
