% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCountTablesGroup.R
\name{MakeCountTablesGroup}
\alias{MakeCountTablesGroup}
\title{Get counts and percentages from a table split by a grouping parameter}
\usage{
MakeCountTablesGroup(data, id_var = "MRN", group_var = "CMD", sum_vars)
}
\arguments{
\item{data}{the input table}

\item{id_var}{the row-identifier column}

\item{group_var}{the parameter to split the calculations by}

\item{sum_vars}{a vector of variables from the table to perform the operation on}
}
\description{
Takes an input table and calculates the total and percentages for each variable per group
In order to ensure one row per observation, a id_var parameter is required as id_var
}
\examples{
out <- MakeCountTablesGroup(data, id_var="MRN", group_var="Included_in_Study", sum_vars=c("Etiology", "Race", "Gender"))
print(out$Etiology)
}
