% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSQLTempTable.R
\name{MakeSQLTempTable}
\alias{MakeSQLTempTable}
\title{Create a Microsoft SQL Server temporary table}
\usage{
MakeSQLTempTable(chan, data, tbl_name, columns, dtypes)
}
\arguments{
\item{chan}{the ODBC channel}

\item{data}{the data frame to copy to SQL}

\item{tbl_name}{the desired temporary table name, without the pound sign passed as a string}

\item{columns}{the columns from data wished to be copied. To be passed as a vector of strings.}

\item{dtypes}{the data types of those columns. Also to be passed as a vector of strings.}
}
\description{
Copy an R data frame into Microsoft SQL Server as a temporary table which can then be used to query other SQL tables
}
\examples{
library(RODBC)

chan <- odbcConnect("AthensSql")
# if pulling data from Bedmaster, all dates should be unix epochs, which is just an integer.
MakeSQLTempTable(chan, data, tbl_name="patients", columns=c("MRN", "Test_Date"), dtypes=c("integer", "integer"))
}
