% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRegress.R
\name{MultiRegress}
\alias{MultiRegress}
\title{Multiple linear or logistinc regression}
\usage{
MultiRegress(table, depen_vars, indep_vars, modtype = "lm")
}
\arguments{
\item{table}{the input data}

\item{depen_vars}{a vector of dependent variables}

\item{indep_vars}{a vector of independent variables}

\item{modtype}{The type of model. Can either be "lm" or "glm"}
}
\description{
Perform a multiple linear or logistic regression (i.e., one with multiple dependent variables)
}
\examples{
out <- MultiRegress(table, depen_vars=c("mRS_Discharge", "mRS_3Month", "TICS_3Month"),
                           indep_vars=c("Total.EEG.Time", "Suppression", "Burst.Suppression"), modtype="glm")
}
