% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizePath.R
\name{NormalizePath}
\alias{NormalizePath}
\title{Normalize Path}
\usage{
NormalizePath(path = utils::readClipboard())
}
\arguments{
\item{path}{file path to normalize.}
}
\value{
converted path, will also write the path to the clipboard.
}
\description{
Converts a file path to always use forward slashes and never
have multiple or a trailing slash.
}
