% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationTest.R
\name{PermutationTest}
\alias{PermutationTest}
\title{Run a permutation test}
\usage{
PermutationTest(group1dat, group2dat, num_permutations,
  test_type = "mean")
}
\arguments{
\item{group1dat}{data from group 1}

\item{group2dat}{data from group 2}

\item{num_permutations}{the number of permutations required}

\item{test_type}{the desired test statistic. defaults to mean.}
}
\description{
Old R function from Biometry that performs a permutation test on a dataset, first split into two by the grouping variable
because THAT'S the most efficient way to do this.
}
\examples{
out <- PermutationTest(group1dat=MaleData, group2dat=FemaleData, num_permutations=500, test_type="mean")
}
