% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PkgInstall.R
\name{PkgInstall}
\alias{PkgInstall}
\title{Install packages}
\usage{
PkgInstall(pkg, upgrade = TRUE)
}
\arguments{
\item{pkg}{Package names or package references. E.g.
\itemize{
\item \code{ggplot2}: package from CRAN, Bioconductor or a CRAN-like repository
in general,
\item \code{tidyverse/ggplot2}: package from GitHub,
\item \code{tidyverse/ggplot2@v3.4.0}: package from GitHub tag or branch,
\item \verb{https://examples.com/.../ggplot2_3.3.6.tar.gz}: package from URL,
\item \code{.}: package in the current working directory.
}

See "\link[pak]{Package sources}" for more details.}

\item{upgrade}{When \code{FALSE}, the default, pak does the minimum amount
of work to give you the latest version(s) of \code{pkg}. It will only upgrade
dependent packages if \code{pkg}, or one of their dependencies explicitly
require a higher version than what you currently have. It will also
prefer a binary package over to source package, even it the binary
package is older.

When \code{upgrade = TRUE}, pak will ensure that you have the latest
version(s) of \code{pkg} and all their dependencies.}
}
\value{
(Invisibly) A data frame with information about the installed
package(s).
}
\description{
This calls \code{\link[pak:pkg_install]{pak::pkg_install()}} which installs a package. This
always clears the caches and has the \code{upgrade} defualted to \code{TRUE}.
}
