% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processREDCapData.R
\name{processREDCapData}
\alias{processREDCapData}
\title{Convert REDCap structured data into something actually usable}
\usage{
processREDCapData(data, redcap_blank = "")
}
\arguments{
\item{data}{the raw data from redcap}

\item{redcap_blank}{the value for any missing data in redcap. defaults to ''}
}
\description{
Convert REDCap structured data into something actually usable
}
\examples{

newdat <- processREDcapData(data=rawredcap, redcap_blank='')
}
