% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeDemographics.R
\name{QuantitativeDemographics}
\alias{QuantitativeDemographics}
\title{Summary statistics on quantitative variables}
\usage{
QuantitativeDemographics(data, id_var, sum_vars)
}
\arguments{
\item{data}{the input data}

\item{id_var}{the row identifier variable--ensures one row per observation}

\item{sum_vars}{the variables to perform calculations on}
}
\description{
Calculates mean, st.dev, median, IQR, min, and max on a vector of variables.
}
\examples{
out <- QuantitativeDemographics(data, id_var="MRN", sum_vars=c("Age", "Admission_GCS", "GOSE"))
print(out$Age)
}
