% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeDemographicsGroup.R
\name{QuantitativeDemographicsGroup}
\alias{QuantitativeDemographicsGroup}
\title{Summary statistics on quantitative variables per group}
\usage{
QuantitativeDemographicsGroup(
  data,
  id_var = "MRN",
  group_var = "CMD",
  sum_vars
)
}
\arguments{
\item{data}{the input data}

\item{id_var}{the row identifier variable--ensures one row per observation}

\item{group_var}{the grouping variable}

\item{sum_vars}{the variables to perform calculations on}
}
\description{
Calculates mean, st.dev, median, IQR, min, and max on a vector of variables per group
}
\examples{
out <- QuantitativeDemographicsGroup(data, id_var="MRN", group_var="CMD", sum_vars=c("Age", "Admission_GCS", "GOSE"))
print(out$Age)
}
