% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeStatistics.R
\name{QuantitativeStatistics}
\alias{QuantitativeStatistics}
\title{Comparative stats on quantitative variables per group}
\usage{
QuantitativeStatistics(data, group_var, tst_vars, test_type = "t.test")
}
\arguments{
\item{data}{the input data}

\item{group_var}{the variable to group by}

\item{tst_vars}{varaibles to analyze}

\item{test_type}{The desired test. Can be either "t.test" or "wilcox.test". Defaults to "t.test"}
}
\description{
Performs either a t-test or wilcox test on quantitative variables per group.
This requires that "MRN" is a column already present in the input data. If this is used in the future,
I might want to change that.
}
\examples{
out <- QuantitativeStatistics(data, group_var="CMD", tst_vars=c("Age", "Hosp_Stay_Len"), test_type="wilcox.test")
}
