% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDB.R
\name{readDB}
\alias{readDB}
\title{Read a multi-sheet excel table into a list}
\usage{
readDB(filepath)
}
\arguments{
\item{filepath}{the filepath of the excel file}
}
\description{
Takes what is essentially an excel database and reads all of the tables (i.e., sheets)
into a list.
}
\examples{
path <- "path/to/file.xlsx"
out <- readDB(filepath=path)
}
