% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExtendedExcel.R
\name{readExtendedExcel}
\alias{readExtendedExcel}
\title{Read an "extended" excel file}
\usage{
readExtendedExcel(datfiles)
}
\arguments{
\item{datfiles}{a vector of all of the filenames to be read}
}
\description{
This will read in an excel file that has its data extend onto multiple sheets.
i.e., it's just one huge table that happens to be split up onto multiple sheets.
...This hasn't ever been tested as a function before.
}
