% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothTimeSeries.R
\name{SmoothTimeSeries}
\alias{SmoothTimeSeries}
\title{Clean a time series}
\usage{
SmoothTimeSeries(data, id_col, var_col, time_col, median_mins, mean_mins,
  n_sec_per_row = 5)
}
\arguments{
\item{data}{the input data frame}

\item{id_col}{name of column, passed as a string, to identify observations by, e.g., a patient identifier}

\item{var_col}{name of variable to perform the smoothing on passed as a string}

\item{time_col}{the column containing the timestamps passed as a string}

\item{median_mins}{window length of data in minutes to perform the median filtering on}

\item{mean_mins}{window length of data in minutes to perform the mean filtering on}

\item{n_sec_per_row}{Length of time in seconds that each row of data corresponds to. Defaults to 5 for BedMaster data.}
}
\description{
Smooth out time series data using a median and then a mean filter.
}
\examples{
## Doing some pre-cleaning of an arterial blood pressure data set before
## piping to the function
tmp <- AllBP_data \%>\%
  filter(ID == id) \%>\%
  group_by(ParTime) \%>\%
  # if multiple measurements at the same time, take the higher one.
  filter(Systolic == max(Systolic)) \%>\%

  SmoothTimeSeries(., id_val = ID, var_col = "Systolic", time_col = "ParTime", median_mins = 30, mean_mins = 30)
}
