% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceEnv.R
\name{sourceEnv}
\alias{sourceEnv}
\title{source the most recent file located in a given directory path.
this function is meant to be used internally within the loadRedcap function.}
\usage{
sourceEnv(path, env)
}
\arguments{
\item{path}{a given path, or vector of paths, conatining a specific file or files to source}

\item{env}{a new environment to source the file into. this is internally created in loadRedcap with new.env()}
}
\description{
source the most recent file located in a given directory path.
this function is meant to be used internally within the loadRedcap function.
}
