% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Try.R
\name{Try}
\alias{Try}
\title{Try an Expression Allowing Error Recovery}
\usage{
Try(expr, otherwise = NULL)
}
\arguments{
\item{expr}{an \code{R} expression to try.}

\item{otherwise}{Default value to use when an error occurs.}
}
\value{
The value of the expression if \code{expr} is evaluated without error;
otherwise \code{otherwise} will be returned.
}
\description{
\code{try} is a wrapper to run an expression that might fail and
allow the user's code to handle error-recovery.
}
