% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateRegress.R
\name{UnivariateRegress}
\alias{UnivariateRegress}
\title{Perform a univariate regression analysis}
\usage{
UnivariateRegress(
  table,
  depen_vars,
  indep_vars,
  mod_type = "glm",
  rand_effect = NULL
)
}
\arguments{
\item{table}{the input data}

\item{depen_vars}{the dependent variables}

\item{indep_vars}{the independent variables}

\item{mod_type}{Type of model to run. Must either be "glm" or "glmer"}

\item{rand_effect}{If mod_type is "glmer", pass whichever variable you want as the random effect as a string. Defaults to NULL.}
}
\description{
Will perform a logistic regression for each independent vatiable in the input table and output the results
in a data frame
}
\examples{
out <- UnivariateRegress(table, depen_vars="Alive_Dch", indep_vars=c("Total.EEG.Time", "Supression", "Beta.Burst"))
}
