% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDescDate.R
\name{UpdateDescDate}
\alias{UpdateDescDate}
\title{Update Description Date}
\usage{
UpdateDescDate(path)
}
\arguments{
\item{path}{Path to a package.}
}
\description{
Updates the Date field in the package description file to the
  current date. Useful for local packages that you want to date which won't
  be uploaded to CRAN.
}
