% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDescVersion.R
\name{UpdateDescVersion}
\alias{UpdateDescVersion}
\title{Update Description Version}
\usage{
UpdateDescVersion(path)
}
\arguments{
\item{path}{Path to a package.}
}
\description{
Updates the version field in the package description file to a
  new version.
}
