% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindowsPath.R
\name{WindowsPath}
\alias{WindowsPath}
\title{WindowsPath}
\usage{
WindowsPath(path = readClipboard())
}
\arguments{
\item{path}{A file path with back slashes to be converted. Default uses the
currently copied text within the clipboard.}
}
\value{
Writes the new converted path to the clipboard and also returns the
  converted path.
}
\description{
Converts a traditional Windows file path with back slashes to
  forward slashes that can be used within R.
}
