% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithoutNamespace_linter.R
\name{WithoutNamespace_linter}
\alias{WithoutNamespace_linter}
\title{Without Namespace Linter}
\usage{
WithoutNamespace_linter(excludePkg = NULL)
}
\arguments{
\item{excludePkg}{Character string, all functions within the namespace will
be excluded from being flagged.}
}
\description{
Check for missing namespace prefixes in function calls.
}
\examples{
# Will produce lints
lintr::lint(
  text = "map()",
  linters = WithoutNamespace_linter()
)

# Okay
lintr::lint(
  text = "purrr::map()",
  linters = WithoutNamespace_linter()
)

}
