% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothTimeSeries.R
\name{SmoothTimeSeries}
\alias{SmoothTimeSeries}
\title{Clean a time series}
\usage{
SmoothTimeSeries(data, id_col = NULL, param_col, var_col, time_col,
  median_mins = NULL, mean_mins = NULL, n_sec_per_row = 5,
  k1 = NULL, k2 = NULL)
}
\arguments{
\item{data}{the input data frame}

\item{id_col}{name of column, passed as a string, to identify observations by, e.g., a patient identifier}

\item{param_col}{name of column, as string, specifying names of parameters to perform smoothing for.}

\item{var_col}{string name of column containing the values for the parameters specified in param_col}

\item{time_col}{the column containing the timestamps passed as a string}

\item{median_mins}{window length of data in minutes to perform the median filtering on}

\item{mean_mins}{window length of data in minutes to perform the mean filtering on}

\item{n_sec_per_row}{Length of time in seconds that each row of data corresponds to. Defaults to 5 for BedMaster data.}

\item{k1}{number of rows to perform median filter on. Only used if the same amount of time does not pass between each row and median_mins is not specified}

\item{k2}{number of rows to perform mean filter on. Only used if the same amount of time does not pass between each row and mean_mins is not specified}
}
\description{
Smooth out time series data using a median and then a mean filter over specified windows of time.
}
\examples{
median_mins = 3
mean_mins = 5

tstdat2 <- SmoothTimeSeries(data = tstdat,
                            id_col = "PATIENT_ID",
                            param_col = "ParName",
                            var_col = "ParValue",
                            time_col = "ParTime",
                            median_mins = median_mins,
                            mean_mins = mean_mins)
}
