% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothTimeSeries.R
\name{SmoothTimeSeries}
\alias{SmoothTimeSeries}
\title{Clean a time series}
\usage{
SmoothTimeSeries(data, var_col, time_col, k1 = 3, k2 = 12)
}
\arguments{
\item{data}{the input data frame}

\item{var_col}{name of variable to perform the smoothing on}

\item{time_col}{the column containing the timestamps}

\item{k1}{number of rows to perform the median filtering on}

\item{k2}{number of rows to perform the mean filtering on}
}
\description{
Smooth out time series data using a median and then a mean filter.
}
\examples{
## Doing some pre-cleaning of an arterial blood pressure data set before
## piping to the function
tmp <- AllBP_data \%>\%
  filter(ID == id) \%>\%
  group_by(ParTime) \%>\%
  # if multiple measurements at the same time, take the higher one.
  filter(Systolic == max(Systolic)) \%>\%
  
  SmoothTimeSeries(., "Systolic", "ParTime", k1 = 3, k2 = 12)
}
